topic_browser: working on my machine
topic_browser.R now doesn't execute anything, it just defines the
function topic_browser(), which Ted should be able to just call without
any parameters. I had to make some changes to the path-munging in
various places to get it to find all the files. I added parameters,
also, to skip the two interactive file-choosing steps.

Fixed a bug in invoking "TopicBulk" for "TopicSize"; the two variables
are identical, but the former is not saved when the aggregate topic
counts are stored.

Added a #quit command so that topic_browser() can exit and restore the
previous working directory.