fix an invisi-bug
python n00b error: didn't convert from string to int. Amazingly this
does not affect the output (since everything was done with dicts AND the
order of lines didn't matter anyway). The only error was that the
last_doc = 0 condition is immediately violated when the first doc was
"0" instead; but the tally dict was empty, so nothing was printed. To
err is Rish, to forgive python-ine.