add SE ribbon to our_geom_smooth
which should be used in all figures, though it's pretty hard to see in
the little faceted plots

also: TODO: ok to use stopped corpus for those word frequency counts?
The text talks about the multiple meanings of "one" but that's not a
word actually counted on the figure currently generated. Wouldn't be
hard to create a term-year matrix for the unstopped corpus instead, if
necessary. Would be surprised if it changed the shape of any curves...