alt_fig_power: uses same x limits
in order to do this, had to tweak the code to keep the original dates
instead of just using the numeric index of the yearly series. Also
removed the part that walks all of the tytm/xxx.rda files to sum
them up. This is equal to the difference between the word-in-topic
per-year counts we are looking at and the total yearly frequency of the
word. The resulting figure looks the same.